[#38] refactor : Change absolute path to relative path
