+++ app.json
@@ -16,7 +16,8 @@
     },
     "assetBundlePatterns": ["**/*"],
     "ios": {
-      "supportsTablet": true
+      "supportsTablet": true,
+      "bundleIdentifier": "com.mokcha.coffee-life-balance"
     },
     "android": {
       "adaptiveIcon": {
