+++ package.json
@@ -1,9 +1,8 @@
 {
-  "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
+    "android": "expo run-android",
+    "ios": "expo run-ios",
     "web": "expo start --web",
     "eject": "expo eject",
     "typecheck": "tsc",
@@ -42,7 +41,7 @@
     "formik": "^2.2.9",
     "react": "16.13.1",
     "react-dom": "16.13.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
+    "react-native": "~0.63.4",
     "react-native-calendars": "^1.1266.0",
     "react-native-elements": "^3.4.2",
     "react-native-gesture-handler": "~1.10.2",
@@ -56,7 +55,10 @@
     "react-native-snap-carousel": "^3.9.1",
     "react-native-svg": "12.1.1",
     "react-native-web": "~0.13.12",
-    "yup": "^0.32.9"
+    "yup": "^0.32.9",
+    "expo-status-bar": "~1.0.4",
+    "expo-updates": "~0.8.1",
+    "react-native-unimodules": "0.14.10"
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
@@ -75,5 +77,7 @@
     "prettier": "2.3.2",
     "typescript": "~4.0.0"
   },
-  "private": true
+  "private": true,
+  "name": "corabal-beta",
+  "version": "1.0.0"
 }
