+++ src/components/atoms/index.ts
@@ -1,0 +1,2 @@
+export { default as CoinIcon } from "./icons/coinIcon";
+export { default as RightArrowIcon } from "./icons/rightArrowIcon";
