+++ src/components/index.ts
@@ -12,14 +12,14 @@ export { default as Calendar } from "./calendar/calendar";
 export { default as RecordList } from "./record-list/record-list";
 export { default as Header } from "./header/header";
 export { default as MainHeader } from "./header/mainHeader";
-export { default as Status } from "./organisms/main/status";
-export { default as Today } from "./organisms/main/today";
-export { default as Month } from "./organisms/main/month";
-export { default as FloatingButton } from "./molecules/main/floatingButton";
-export { default as Banner } from "./molecules/main/banner";
-export { default as CircleProgress } from "./molecules/main/circleProgress";
-export { default as Aim } from "./molecules/main/aim";
-export { default as RecordsOfMonth } from "./molecules/main/recordsOfMonth";
-export { default as Point } from "./molecules/main/point";
-export { default as RightArrowIcon } from "./atoms/icons/rightArrowIcon";
-export { default as CoinIcon } from "./atoms/icons/coinIcon";
+// export { default as Status } from "./organisms/main/status";
+// export { default as Today } from "./organisms/main/today";
+// export { default as Month } from "./organisms/main/month";
+// export { default as FloatingButton } from "./molecules/main/floatingButton";
+// export { default as Banner } from "./molecules/main/banner";
+// export { default as CircleProgress } from "./molecules/main/circleProgress";
+// export { default as Aim } from "./molecules/main/aim";
+// export { default as RecordsOfMonth } from "./molecules/main/recordsOfMonth";
+// export { default as Point } from "./molecules/main/point";
+// export { default as RightArrowIcon } from "./atoms/icons/rightArrowIcon";
+// export { default as CoinIcon } from "./atoms/icons/coinIcon";
