+++ src/components/molecules/index.ts
@@ -1,0 +1,6 @@
+export { default as Aim } from "./main/aim";
+export { default as Banner } from "./main/banner";
+export { default as CircleProgress } from "./main/circleProgress";
+export { default as FloatingButton } from "./main/floatingButton";
+export { default as Point } from "./main/point";
+export { default as RecordsOfMonth } from "./main/recordsOfMonth";
