+++ src/components/organisms/index.ts
@@ -1,0 +1,3 @@
+export { default as Month } from "./main/month";
+export { default as Status } from "./main/status";
+export { default as Today } from "./main/today";
