+++ src/pages/signUp/templates/signUp.style.ts
@@ -1,0 +1,25 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: "center",
+    justifyContent: "center"
+  },
+  emailContainer: {
+    marginBottom: 41
+  },
+  linkContainer: {
+    flex: 1,
+    alignItems: "center",
+    justifyContent: "center",
+    marginBottom: 53
+  },
+  LinkText: {
+    fontSize: 14,
+    lineHeight: 20,
+    color: "#757474"
+  }
+});
+
+export default styles;
